requirements.txt
flask==3.0.3
flask-cors==4.0.1
rapidfuzz==3.9.6
Pillow==10.4.0
openai==1.43.0
python-dotenv==1.0.1
pytrends==4.9.2
